citus added

+++ app/models/todo.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Integer, String
+from sqlalchemy import BigInteger, Boolean, Column, String
 
 from db import Base
 
@@ -6,7 +6,7 @@ from db import Base
 class Todo(Base):
     __tablename__ = "todos"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     title = Column(String(200), nullable=False)
     body = Column(String(2000), nullable=False)
     done = Column(Boolean, default=False, nullable=False)
